move `prompt-options.js` to utils/input dir
