+++ bin/now-deploy.js
@@ -22,7 +22,6 @@ const { version } = require('../lib/pkg')
 const Logger = require('../lib/build-logger')
 const Now = require('../lib')
 const toHumanPath = require('../lib/utils/to-human-path')
-const promptOptions = require('../lib/utils/prompt-options')
 const { handleError, error } = require('../lib/error')
 const { fromGit, isRepoPath, gitPathParts } = require('../lib/git')
 const readMetaData = require('../lib/read-metadata')
@@ -35,6 +34,7 @@ const info = require('../lib/utils/output/info')
 const wait = require('../lib/utils/output/wait')
 const NowPlans = require('../lib/plans')
 const promptBool = require('../lib/utils/input/prompt-bool')
+const promptOptions = require('../lib/utils/input/prompt-options')
 const note = require('../lib/utils/output/note')
 
 const argv = minimist(process.argv.slice(2), {
