+++ lib/utils/input/prompt-options.js
